производит модуляцию
